Added mini image also to main program

+++ SLM_controller.py
@@ -199,7 +199,7 @@ class TkinterDisplay:
             mini_image[-1,0:-1,:] = 255  # Right edge
             mini_image[0:-1,-1,:] = 255  # Bottom edge
 
-        self.mini_image = mini_image
+        self.mini_image = mini_image.astype('uint8')
 
     def create_algorithm_selection(self, x_pos, y_pos):
         self.selected_algorithm = StringVar()
@@ -367,7 +367,7 @@ class TkinterDisplay:
          self.create_trap_image(trap_x=c_p['traps_absolute_pos'][0],
             trap_y=c_p['traps_absolute_pos'][1])
 
-         self.mini_image = self.mini_image.astype('uint8')
+         #self.mini_image = self.mini_image.astype('uint8')
 
          self.photo = PIL.ImageTk.PhotoImage(image = PIL.Image.fromarray(self.mini_image, mode='RGB'))
          self.canvas.create_image(0, 0, image = self.photo, anchor = tkinter.NW) # need to use a compatible image type
